Install Pact for contract testing
We’ve decided to try this out as a way of defining the intervention
service API contract, and also as a way of mocking out the intervention
service in our client class’s unit tests.

We initially tried using the plain pact package, but we encountered some
timeouts in the Jest tests, and saw the jest-pact package recommended as
a way of resolving these. So we’ve tried that out and it seems to work.

Co-authored-by: Lawrence Forooghian <lawrence@dxw.com>