Inject an InterventionsService into the routes
We’ll be using this as we start to add controller actions which load and
create referrals.

This also refactors the way that the service mocks are set up in the
test, to give us the ability to inject a mock into the test app.

We need easy access to the mocks instead of them being buried away
inside the app, so that we can make assertions on which methods have
been called etc. Also we'll probably want to be able to configure
them to return custom responses at some point.

This pattern is copied from the use-of-force repo [1].

[1] https://github.com/ministryofjustice/use-of-force/blob/92d13e58f35755819aa37a1fb6d2c3ca7bb444b4/server/routes/testutils/appSetup.ts#L62

Co-authored-by: George Eaton <george@dxw.com>