Add start referral page and an empty referral form
This is our first piece of interventions-specific user-facing
functionality. The user can click “Start now”, which creates a referral
on the interventions service and then redirects them to an empty
referral form. The referral form is currently just a placeholder that
replays the referral’s ID. We’ll build it out properly in IC-613.

The start page design is taken from our alpha prototype.

We haven’t added any integration tests since we haven’t yet decided an
approach for mocking. This is something we’ll address in IC-627.

As in 506d5c6, the approach for mocking services in the controller tests
is taken from use-of-force.

We tested this locally using the Pact stub server. You can run it with

```
docker run -t -p 8092:8080 -v "$(pwd)/pact/pacts/:/app/pacts" pactfoundation/pact-stub-server -p 8080 -d pacts
```

Co-authored-by: George Eaton <george@dxw.com>