Enable the noImplicitAny TypeScript config
Given that we’ve chosen to use a strongly-typed, compile-time checked
language, we should make the most of what this gives us. My ultimate aim
is to turn on the "strict" compiler option, which turns on a whole bunch
of safety options. But I’m going to do them one option at a time so that
I can do a little bit on it each time I have some downtime.

We start with turning on noImplicitAny, which forces us to be a lot more
explicit about our types throughout the codebase, but increases our type
safety and ease of understanding for a developer looking at the code.