Display a task list on the referral form
This displays a task list which currently shows a single fixed state
- the user has half filled the referral form, and the next thing for
them to do is to fill in the desired completion date for the
accommodation intervention.

The markup and styling here is taken from the corresponding screen in
the prototype.

Architecturally, here I’ve started to think about the different
responsibilities that exist within the app. I’ve introduced the concept
of a presenter, which is responsible for converting model data into
structured data that’s ready to be directly presented by a view. For
example, in the case of the referral form, it would be responsible for
populating the task list sections, and deciding each of their statuses
and enabled-ness based on the data in the referral. It’d also be
reponsible for things like formatting model data.

The main aim of the presenter is to keep the view templates as simple as
possible, so that they're basically just looping over things and
rendering pre-formatted strings. For example, even though none of the
business logic for driving the referral form has yet been implemented,
the view template will not need to change at all as we start to
implement that logic - it’ll just be a change to logic in the presenter.

The presenter is easier to test than a view template, too.

On a practical note (separation of responibilities aside), we found in
the prototypes that writing any kind of logic in the view templates is a
deeply unpleasant experience, because Nunjucks is a rather difficult
language to work with compared to e.g. TypeScript.

I’ve also introduced the concept of a “view”, which is a class that
bundles up a Nunjucks template name and some helper functions, and can
be passed directly to a `render` call in the controller. Not sure how
useful this one will be, let’s see.