Add a patchDraftReferral method to interventionsService
Keen for feedback on whether a PATCH or a PUT is more appropriate here.

My thinking was that we need some way to handle the fact that some of
the field validations are context-dependent - for example we only need
to validate the presence of a field when the user is actually filling in
that particular field in the UI. It’s okay for that field to be null
before that point, and the backend validations shouldn’t forbid it. I
thought that using a PATCH request would allow the frontend to tell the
backend which fields are actually being updated and hence which ones it
should apply the validations to. But perhaps there are other ways of
handling this.