Reduce initial delay for quicker deployment time
A deployment for the UI always takes over one minute, however, the pods
are up within seconds:

```
$ k get pod/hmpps-interventions-ui-7c7695fd9b-mkvd9  -ojson | jq '.status | .startTime, .containerStatuses[].state.running.startedAt'
"2020-12-07T09:33:30Z"
"2020-12-07T09:33:37Z"
```

To avoid stalling the deploy for 60 seconds (the value of the
`readinessProbe.initialDelaySeconds`) we can set it to 10 seconds.

I feel the default 10 seconds value for `periodSeconds` also makes
sense.