Use TCP port scan for probes instead of /ping
The `/ping` endpoint does not exist on this application.

I would use `/health` but it checks downstream services, and while it's
great to check the overall health of the application, it shouldn't be
used to determine whether to route traffic to it or when to restart.

From the TCP liveness probe documentation:

> With this configuration, the kubelet will attempt to open a socket to
> your container on the specified port. If it can establish a connection,
> the container is considered healthy, if it can't it is considered a
> failure.

Source: https://kubernetes.io/docs/tasks/configure-pod-container/configure-liveness-readiness-startup-probes/#define-a-tcp-liveness-probe