Add Pact verification against 'main' tag consumers
Since the broker is open for reads, this test can run without
authentication.

Currently, it does not publish its verification results.

The verification logs neat stuff during the test run, which is hidden by
the test runner; `./gradlew check --info` shows it, but it's awkward. Is
there a way to turn logging on just for this test file?