[to revisit] Disable authorisation for test clients
I'm not familiar with Spring Boot :)

All Pact tests fail with
`1.1) status: expected status of 200 but was 401` because all of the
requests are unauthorised.

I wanted to either

1. create a state in the Pact test that all incoming requests are
   pre-authorised
2. or remove the authentication filter from the request filter chain

I couldn't do either of these tidily -- the other examples in the
organisation all involve significant amount of extra code which I wanted
to avoid.

So for now, this is a workaround: the permission checks only run on
non-test profiles.