Add the express-validator package
As part of allowing the user to set the completion deadline on a draft
referral, I’ll need to validate the data that they input. I had
originally thought about trying to do it myself but I thought that this
is perhaps a solved problem, so I looked for some libraries. This is the
one that seems to be mentioned the most. I’ll try it out and see what I
think.

Co-authored-by: George Eaton <george@dxw.com>