Create form class to handle deadline input
The idea of this form class is that it handles all of the user input,
including:

- validation
- sanitization
- returning a list of parameters that can be directly passed to the
  InterventionsService#patchDraftReferral method

The express-validator recommended way of adding validation is to do it
in a middleware, but with our current setup that would have meant having
to add it in server/routes/index.ts, which would have quickly got out of
hand. I like having it contained in a class. I’m not thrilled about the
fact that this mutates the request, but will live with it for now.

I don’t know whether we’ll end up with one form class per screen, or
(more likely, I think) a configurable form class that can handle any
kind of update to the draft referral.

There’s some work remaining to figure out how to meet the error handling
requirements of the GOV.UK Design System, which specifies quite a
granular level of error handling for date inputs. I decided not to get
into that here because it seemed like an absolute rabbit hole.

Tom suggested that we add the error messages into a separate file so
that they could be easily reviewed by designers.

Co-authored-by: George Eaton <george@dxw.com>