Add completion deadline presenter
The responsibilities of this presenter are the same as those in the one
introduced in bbf8c18 – format data ready for a view to render. But here
it’s more complex, because this presenter is being used to drive two
different states (pre-submission and post-submission with errors) so the
data can come from either a referral object, or user input, and there
are also potentially errors to be displayed.

Co-authored-by: George Eaton <george@dxw.com>