Add completion deadline view
I’ve made heavy use of the Nunjucks macros provided by the GOV.UK
Frontend in this template. It makes the templates super-simple. I think
that since we’re using very standard components (date input, error
summary) we’re unlikely to want to customise the markup and hence safe
to use the macros.

We use the view class as a place to prepare the arguments for the
various Nunjucks macros. This is much nicer than trying to do it inside
the template.

Co-authored-by: George Eaton <george@dxw.com>