Verify Pact consumer contracts and publish results on CI
Done via an explicit Gradle task that converts the environment variables
to system properties expected by the `@PactBroker` annotation and the
`PactBrokerLoader` class.

The annotation contains a value resolver that attempts to retrieve the
value from environment variables, however, the resolution did not seem
to happen for me.

While ugly, the hand-wiring of env-var to system-property works and at
least it's explicit (even if the property usage is hidden in
`@PactBroker`)