Add 'completionDeadline' field to DraftReferralDTO
This change also makes a subtle change to how the DTO is constructed,
allowing the serialization of the fields to take place only when the DTO
is converted to JSON for a response.

This change also includes a test which checks the DTO can be used to
deserialize partial referral updates.