Add needs and requirements view
This is larger than previous views. Partly because it has more fields,
but it also has some new things, like conditional radios (where we
define the text inputs’ HTML in the template and pass it in, so that we
can easily use the Nunjucks macros to generate the inputs), a summary
list, escaping HTML content for summary lists.