Add the Fishery library
We’re at the point where it’s getting hard to keep all of our
DraftReferral test instances updated as we add new properties to the
model. In Ruby-land, we’ve quite liked using factory_bot for this in the
past. Fishery is a similar thing for TypeScript, by the makers of
factory_bot [1]. Let’s give it a go — it’s pretty new so it might have
some teething issues or might become unsupported, but I imagine we could
replace it if need be, and it's only a dev dependency.

[1] https://thoughtbot.com/blog/announcing-fishery-a-javascript-and-typescript-factory-library