Add needs and requirements form
Some of the error messages here contain information interpolated from
the referral. I’m not sure whether it makes more sense for these to be
interpolated by the form or the presenter - it maybe is presentation
logic. But it was easier to do it in the form so that’s what I’ve done.
Let’s see how that goes and whether we find ourselves needing to do
anything more complex than just needing to stick a model value directly
into a message.