Change `ServiceCategory` structure and endpoint
We previously had a `serviceCategory` object on the `DraftReferral` that
had its own ID, and then would make a separate request for the
`complexityLevels` for that `serviceCategory`. Now we're adding more
fields to the `serviceCategory`, it feels as though we might want it to
be a more fully-fledged object, and to store a reference to its `id` on
the `Referral` instead.

Now, a request to `service-category/:id` will return:
```
  {
    id: 'uuid',
    name: 'accommodation',
    complexityLevels: [
     {
       'id': 'uuid',
       'title': 'low complexity',
       'description': 'complexity level description',
     }
    ]
  }
```

The `referral` has a `serviceCategoryId` instead of the whole service
category.

This does mean we're making more requests to the API for both entities,
so we might want to consider the performance implications of this.

For further context, see:
https://github.com/ministryofjustice/hmpps-interventions-ui/pull/44#discussion_r543222831