Add a DraftReferral factory
The idea is that we’ll be able to use this whenever we want to create a
draft referral in our tests, without needing to each specify a bunch of
properties that are irrelevant to the test in question.

We add the factory files to the the list of files that ESLint’s
import/no-extraneous-dependencies rule allows to contain imports of
modules from devDependencies.