Enable bodyParser urlencoding in tests
This allows us to submit a HTTP form with repeated values for a given
key, and access them as a single array in the Express request’s `body`
object.

This allows us to send checkbox values in requests and access them in
the body as an array like `desired-outcomes-ids` (rather than
`desired-outcomes-ids[0]` and `desired-outcomes-ids[1]` as was
happening when I tried to access these values).

I'm not quite sure why it's manually disabled, this may be a problem
with the HMPPS Typescript project that we used as a template.

Co-authored-by: Lawrence Forooghian <lawrence@dxw.com>
(Addressing code review comment in George’s absence)