Turn on strictNullChecks TypeScript check
This makes the TypeScript type system aware of null and undefined and
helps us catch a bunch of errors that might arise from not accounting
for one of these values.

I’ve made use of non-null assertions [1] sparingly, in the following
cases:

- where, if the operand is null, it’s a sign of a bigger problem that
  will get caught anyway (e.g. in unit tests, the assertions will fail
  anyway or we’ll get a runtime error in the tests, either is fine)

- where I’m pretty sure that the operand will not be null in the way our
  app is configured (request.session) - this is a shakier one but we can
  tighten it up later maybe

- where additional runtime information that the compiler doesn’t know
  about tells me that the operand is non-null — e.g. I already checked a
  (string | null) value is non-blank

I’ve turned off the ESLint rule that forbids non-null assertions. If we
think this too lenient then we can tweak it per-file / per-line, but I
think we can see how often these sorts of issues come up in code review
and take it from there.

[1] https://www.typescriptlang.org/docs/handbook/release-notes/typescript-2-0.html#non-null-assertion-operator