Treat ESLint warnings as errors
I think that any code quality tools should cause CI to fail when they
find a problem. Warnings that can be ignored by developers will, in
general, mount up unnoticed. And it becomes difficult for even a
developer who is paying attention to the ever-growing list of warnings
to figure out if they’ve introduced new ones in their changes.

So now if a warning happens, we fail the linting step. We should handle
a warning either by:

- changing the code to satisfy the rule
- disabling the rule if we disagree with it
- selectively turning the rule off for a particular file (through ESLint
  comments or config), or particular lines (through ESLint comments)