Check already-selected desired outcomes on page
e.g. if a user goes to change their answer.

As before, we haven't really got an integration test for this, as we
haven't figured out how to configure Cypress with this front/back end
split.

This also adds a pact stub for pre-selected desired outcomes to allow us
to test the functionality locally with a referral of id
`037cc90b-beaa-4a32-9ab7-7f79136e1d27` - these are starting to get a bit
unmaintainable, it would be good to find a more convenient solution for
running locally.

Co-authored-by: Lawrence Forooghian <lawrence@dxw.com>
(Addressing code review comment in George’s absence)