Create ReferralDataPresenterUtils
In all our presenters, we find ourselves needing to choose whether to
replay the model data or the data that the user input, and then
converting things like 'yes' / 'no' to booleans. This was going to
become very tedious to test in each of our presenter tests, so I’ve
pulled it out into a helper class which can be reused across the
presenters and tested in detail, hopefully allowing us to be a little
more loose in the tests for the individual presenters.