Create FormUtils
We’re going to have some patterns that we repeat in our forms, with a
bunch of fiddly validation logic that we need to test. I really don’t
want to test this every time we write a new form, so I’ve pulled the
validation logic for a yes / no radio button, with an input which is
only validated when the value is yes, into a class. Also a method for
running validations on a request and returning a validation result.