Add service user’s first name to DraftReferral
I need this so that I can display it in some user facing messages. This
is the first time that we’ve thought about service user information.
It’ll probably be pulled from an external system (Delius) in the future,
but I imagine we’ll still be adding that data to the draft referral too.
Not saying that firstName is a good field name or the right thing to use
when displaying a user’s name, though.