Allow users to submit further information
I haven't gone with our usual pattern of using a `Form`-type class here,
as we don't really have any validations to worry about. I thought this
was small enough to do in the controller.

It's worth noting that when there's an empty text area, the
`further-information` value is POSTed as an empty string. This may be a
way of tracking the user's progress (i.e. the value not being `null`),
but there might be another way.