Configure Offender Assessments API
To allow us to display OASys data in our application.

If the interventions UI is configured to use URL `INTERVENTIONS_UI_AUTH_URL` to
communicate with the HMPPS auth service, then when the interventions UI
requests a token from the auth service, the token that is returned will contain
an iss with the same hostname and port number as `INTERVENTIONS_UI_AUTH_URL`.

When the assessments API service, configured to use `ASSESSMENTS_API_AUTH_URL` to
communicate with the HMPPS auth service, is given an authentication token by
the interventions UI, it will check that the iss of this token has hostname and
port number that match those of `ASSESSMENTS_API_AUTH_URL`. It will “fail”
with `Unable to resolve the Configuration with the provided Issuer of
`ASSESSMENTS_API_AUTH_URL`` if they don’t match.

This means that `INTERVENTIONS_UI_AUTH_URL` and `ASSESSMENTS_API_AUTH_URL` need to
have the same hostname and port number.

We run the interventions UI outside of the Docker Compose network, and so the
value that we use for `INTERVENTIONS_UI_AUTH_URL` is `http://localhost:8090` (port
number of hmpps-auth container).

When using Docker Compose to run the assessments API service, the easiest way
to provide the assessments API service with a URL that refers to the auth
service is by giving it an `ASSESSMENTS_API_AUTH_URL` of
`http://hmpps-auth:8090` (internal port number of hmpps-auth container).

This means that, in order to satisfy the required relation between
`INTERVENTIONS_UI_AUTH_URL` and `ASSESSMENTS_API_AUTH_URL`:
- the host port number and the internal port number of the hmpps-auth container
need to be the same (hence the change to the port mapping in the
`docker-compose.yml`)
- the hostname `hmpps-auth` needs to resolve to localhost on the host machine
(this will require a change to our local `/etc/hosts` - instructions for
this are added to the README.)

Co-authored-by: Lawrence Forooghian <lawrence@dxw.com>
Co-authored-by: Tom Myers <tom.myers@digital.justice.gov.uk>