Add offender assessments api service
with sample function to request latest assessement from OASys.

This isn't tested at the moment, and I'm not sure exactly what fields
we'll want to use in our application, but this provides a starting point
so we can at least make requests to the api with our app and inspect the
various responses.