Convert Pact build to be triggerable via CI API
The CircleCI v2 API supports triggering pipelines on a GitHub project.

However, it only supports branch/tag arguments and pipeline parameters:
there is no built-in support for selecting a specific workflow or job.

Instead, we need to use **pipeline parameters**.

These can be passed down from the triggering API call and we need to
implement a branching logic based on it.

Tested the below code with the following API call:
```
$ curl -X POST \
  'https://circleci.com/api/v2/project/github/ministryofjustice/hmpps-interventions-service/pipeline'
  -H "Circle-Token: $CIRCLECI_INTERVENTIONS_TOKEN" \
  -H "Content-Type: application/json" \
  -d @payload.json
```

payload.json:
```
{
  "branch": "trigger-pact-builds",
  "parameters": {
    "only_pacts": true,
    "pact_consumer_tags": "last-implemented"
  }
}
```