require a 'serviceUserCrn' param in create draft referral requests
this ensures that we cannot create a new referral without a user associated with it. the reason for this
is that changes in the journey mean that there will no longer be a section of the referral form to select
a service user. the service user is pre-selected in the 'discover' part of the product, and is known at
refferal creation time.