Add `SerializeDeliusUser` method and Delius fields
This takes a `DeliusUser` and transforms it into a `ServiceUser` that
can be used to patch the `Referral` on the Interventions Service.

I'm not 100% sure the Interventions Service is the best place for this
function, so would appreciate any other suggestions.

Uses `CalendarDay` function to ensure dates are formatted in a
consistent way, like the CompletionDate we currently send back to the
InterventionsService.