Re-enable migration and exception logs
Currently, the migration logs and raised exceptions are not visible
locally or on the development Kubernetes pod logs.

It looks like a lot of messages are redirected to
`ApplicationInsightsAppender` -- which we do not yet use.

I like seeing the migration status and exceptions on the screen, so:
this commit reverts back to default log settings until we configure
AppInsights.