Reduce initial delay for quicker deployment time
Currently, a deploy takes less than 10 seconds:

```
$ k get pod/hmpps-interventions-service-5ff6678c86-f84r7  -ojson | jq '.status | .startTime, .containerStatuses[].state.running.startedAt'
"2021-01-21T20:55:15Z"
"2021-01-21T20:55:21Z"

$ k get pod/hmpps-interventions-service-5ff6678c86-7548w  -ojson | jq '.status | .startTime, .containerStatuses[].state.running.startedAt'
"2021-01-21T20:55:15Z"
"2021-01-21T20:55:23Z"
```

To avoid stalling the deploy for 60 seconds (the value of the
`readinessProbe.initialDelaySeconds`) we can set it to 10 seconds.

I feel the default 10 seconds value for `periodSeconds` also makes
sense.

This change is similar to
https://github.com/ministryofjustice/hmpps-interventions-ui/commit/ab000c6ecaa41f25c620e77732114c8d36ffdba4.