Add methods to format dates in GOV.UK style
I’ve decided to always go via the CalendarDay class, even when we have a
Date instance that we could directly format. Because sometimes we won’t
have a full datetime - e.g. completion deadline, date of birth - and so
we’ll need to be able to format a CalendarDay anyway, so makes sense to
just have a single API for day formatting.