Add routing and form links to Service User Details
The `POST` route accessed when clicking "Save and Continue" isn't really
doing much at the moment, apart from re-routing to the form. In future,
we'll probably want to pass some sort of field back to the API or to the
ReferralForm object to mark the section as completed, but we haven't
decided on the best way to do that just yet.