Pass a fresh object to Fishery’s params each time
After seeing some surprising behaviour, where creating a second instance
of SentReferral using the factory would modify the first instance, I
found out that it looks like Fishery will mutate the `params` object
that you pass it. So we shouldn’t use a shared object.

I think this could be considered a bug in Fishery, though; it probably
should be copying the objects that are passed to it. I’ve created IC-992
for us to investigate.