Add cypress-get-table package
I want an easy way to check the contents of an HTML table. In the
interests of time, which I’ve already spend quite a lot of on this
story, I decided not to spend too long on the Cypress docs and to, for
now, just pull in a package that does this. I would like to get rid of
this package, though, since it only has a single function and I’m hoping
to get better at Cypress to be able to write this sort of thing myself.

Created IC-994 for getting rid of this.