Proxy interventions service requests through Wiremock locally
This adds a local instance of Wiremock to be used when running the
application in development.

It allows us to:

- use the real interventions service for functionality that has been
  implemented

- provide stubbed responses for testing out parts of the UI that depend
  on things not yet implemented on the interventions service

- provide stubbed responses for testing out how the UI behaves with an
  interventions service response that is hard to set up

This hopefully allows us to have the best of both worlds, and will allow
us to stop using the Pact mock server. We should focus on making our
Pact examples be good tests, and stop trying to simultaneously make them
good mocks capable of demonstrating entire journeys in the UI.