Use a separate Redis instance for Cypress tests
Until now, the docker-compose.yml and docker-compose-test.yml have
shared the same service (just named 'redis'). This means that running
`docker-compose down` for the tests would kill the local dev Redis
instance too, which in turn kills the server. This was quite annoying.

Also, it makes sense for the local dev and Cypress environments to be
completely isolated from each other.

The CircleCI port configuration is from [1].

[1] https://github.com/bitnami/bitnami-docker-redis#passing-extra-command-line-flags-to-redis-server-startup