Pass in service category to referral form
Now that the "Find" section is happening before the referral form, this
will always be present on the form. As that's a bit more work to change
now, I'm defaulting to "Accommodation" if a service category id isn't
present on the referral (as it was initially to be set as part of this
user journey).