Configure Wiremock’s prefix for interventions service
In the Cypress tests, Wiremock provides all intervention service
endpoints at /interventions. This is done so that it can also mock other
services, e.g. HMPPS Auth at /auth.

We’d like do to the same for the development Wiremock server, too,
however there is no built-in ability for Wiremock to strip a prefix from
the URL when proxying it. So we need the Wiremock paths to be the same
as the interventions service paths.

Somebody has requested this feature here [1] and it was suggested that
there is an ability to extend Wiremock to build it, but until we need to
mock multiple services there’s no need to think about it.

[1] https://github.com/tomakehurst/wiremock/issues/745 proxyUrlPrefixToRemove