Add ability to specify interventions service mocks in development
We can set up mocks in the mocks.ts file. I’ve given an example of
mocking the /sent-referrals endpoint, which isn’t yet implemented in the
interventions service.

There’s some server code that runs only in the development NODE_ENV,
which sets up the mocks. I did consider instead making this a separate
utility to be run independent of the server, which seemed like a cleaner
thing to do, but if I wanted to write it in TypeScript then I’d have to
find some way of building that code and running it alongside the server
and so it just seemed easier to integrate into our existing build/run
pipeline.

I’m not sure whether we want to commit these mocks or if they should be
a local-only thing. Sharing the mocks between frontend developers is
useful whilst waiting for something to be implemented on the backend,
but then we have to remember to remove them when the functionality gets
implemented. I’ve committed them for now but would welcome thoughts.