Create draftReferral with hardcoded interventionId
This will be hardcoded until we start work on the 'Find' journey, which
will pass the selected `interventionId` to the `#createDraftReferral`
method.

At the moment, we don't actually need to update the response, as the
'Refer' journey doesn't need anything from the intervention. We're
currently getting all the information we need from the service category.

This will allow us to store things like a `location` or its ID on the
intervention when we need to display it in future.