Add a basic service provider referral details page
As part of this sprint’s critical outcome of “a referral can be sent to
a provider”, we wanted to demonstrate that submitting a referral causes
the service provider to receive details of the referring probation
practitioner.

We decided that the best way to do this is to build a basic version of
the referral details page, where we can use the information contained
in the SentReferral object to fetch the probation practitioner record
from Delius and display their details.

We had a brief chat about whether the right thing to display here is the
details of the author of the referral, or the service user's current
probation practitioner. I did a bit of research into how we would fetch
the latter from Delius, and ended up in a rabbit hole of trying to
figure out some of the business concepts and the difference between a
"responsible officer" and an "offender manager" in Community API
vocabulary. Since we don't have any concrete business requirements about
the display of PP data yet, we decided to instead just focus on
displaying the referring PP.

Also, the Community API `/secure/users/${username}/details` endpoint
doesn't return many details about the user, so if we want to display
things like phone number / address then we still need to figure out how
to do that.