Make it easy for designers to build templates
We’ve decided that the interaction designers are going to start helping
us to build some of the MVP templates, where they’ll be especially handy
for some of the more custom components / styling.

To make things as simple as possible for the designers, we’re going to
set up all of the routing and blank templates ahead of time, so that
they just need to fill in the template code.

This adds a mechanism for serving static pages, by adding them to the
`pages` array in StaticContentController, and adds an index page listing
all of the static pages, so that given a Jira issue, the designers
can easily find the template file they need to modify, and can preview
how the page looks in a browser.

The static pages are only enabled in local environments (development and
test).