Add contact details to service user
when starting a referral.

As with other SU details, too many of these can come back as null, or
not even present in the response, so there's a lot of null checks here.

In the 'Refer' journey, we don't need to display these anywhere (they're
only going to be displayed in the `Receive` journey for the Service
Provider, so they haven't been added to the presenter.