Display SU details on Referral page
We've had to hardcode a CRN on the mocks for this, when running locally
with WireMock, so it'll always show Aadland Bertrand locally, which will
lead to a discrepancy with the other service users, but I think this
will do for now.

I've opted to use the Notification banner for now, as this is the
nearest component I could find on the GOV UK design system, but we'll
probably replace it in future with something more bespoke.

Again, this implementation relies on local data from the Community API,
which is currently very lacking. I'd like to add some better examples so
we can see actual contact details.