WIP Filter component
This is mostly a copy of the code written by the Digital Marketplace
team here:
https://github.com/alphagov/digitalmarketplace-govuk-frontend/tree/master/src/digitalmarketplace/components/option-select,
based on:
https://finder-frontend.herokuapp.com/component-guide/option-select.

I've struggled with this for half a day, it's almost there but for some
reason the "closed" behaviour doesn't work. I'd like to dig into why,
but I feel like I've already spent too much time on this.

This code I've copied from the other team is old-style JS, rather than
the ES6 we're using, so it would be best to re-write this ourselves, as
I haven't had time to dig into how it works too much.