Add SP sessions template
This was copied from the standard template that comes with the GOV UK
Frontend (in `/node_modules/govuk-frontend/govuk/template.njk`, also
visible
[here](https://github.com/alphagov/govuk-frontend/blob/master/src/govuk/template.njk)

This is a temporary step, and we'll want to undo it and just extend the
standard GOV UK template, rather than forking it, to make sure we get
any changes to the design system.

This will be used in place of the standard GOV UK template as the SP
session pages have a different layout to the standard design. This
template will be extended by a "layout" that we can import in each page
to ensure they all have the same components, e.g. sidebar, sub nav, SU
details banner.

To use this, you can override the block content in the template, which
will mean we'll need to change the existing templates to use things like
`{% block mainContent %}` rather than simply `{% content %}`