Allow desired outcomes to have their service category set when created in code.
Currently all the desired outcomes in the test database are created by migration scripts.
To create them in code and associate them with a service category was impossible, because
the service category id column was missing from the entity.

This change adds the service category id field, and factory methods for creating and persisting
desired outcomes and service categories in the tests.