Strictly define which initial fields are required to create a referral entity.
This change ensures that referrals cannot be created without the following fields:
 - id
 - createdAt
 - createdBy
 - serviceUserCRN
 - intervention

These fields are now `val`s, with no default values. The auto generation of id and createdAt
timestamps are changed from using annotations, to generating the values ourself. This allows
these values to be overridden in tests.

Note that this commit breaks the tests! The changes to get them fixed are wide reaching, and will be
addressed in subsequent commits.