Intervention search results can be filtered
By PCC region, gender, and age.

The filter UI that I’ve built is very basic. In the future, we might
want to improve this to use one of:

- the MOJ Design System filter component
- a custom design
- the GDS Finder Frontend "option select" component

As for the filters displayed here, I spoke with user researchers. We
agreed that we didn’t explore these filters much in alpha and don’t yet
know much about the user needs. Decisions for first version we’ll test:

- gender filter: we won’t provide a way to filter for gender-exclusive
  interventions through this UI, and instead only allow filtering to
  find interventions suitable for a particular service user

- age filter: we’ll have a single checkbox, that allows them to view
  only the 18-25 interventions