Use real database for all tests.
This commit adds an annotation that we can use to replace 'DataJpaTest'
which ensures the real database is used for all tests. There are two reasons
for doing this. Firstly the in memory H2 db doesn't support the array types
we are using. Secondly it ensures we catch any bugs that may arise from subtle
difference in sql syntax or behaviour.