Add tabbed section to intervention details page
Currently just displaying one tab with the service provider's name -
we'll want to update this to show more contact details when we add them
to the Service Provider model on the backend.

This also adds a `govukFrontendTypes` file that will hold the objects we
pass to Nunjucks Macros. We'll want to add to this and refactor existing
places where we could define similar interfaces more strongly.