Ensure `desired-outcomes-ids` are always an array
Turns out if you don't use an array on the checkbox `name`, `bodyParser`
treats single checkbox selections as a string, rather than an array,
which the backend didn't know how to handle. Before this, when a user
selected a single desired outcome, the backend would return a 500 and
break the user journey.

This commit should be in the running for some sort of award for longest
time spent trying to figure out an issue, solved by adding 2 characters.