Use ID of local intervention for mock intervention
Because of the way our mocks are set up, the Refer journey is using the
real Interventions Service, but the Find journey is using mocked
journey. This means we go from using mocked data to real data, so we
need to ensure we're at least using the ID of an existing intervention
when we get to the Refer journey.

This is obviously a temporary thing, and won't matter when the backend
catches up and we stop mocking it in the Find journey.