Safety: impose upper limit on number of tries
With an infinite sequence + find, there is a potential that we will have
infinite loops if a) everything's used up b) there's a bug in the
future.

To avoid a request thread completely blocking, limit the number of tries
in getting a unique reference number to 10.

There is no science behind the choice, it seemed OK to start with 10.

The change also refactors to a more conventional control flow.