Use correct AuthUser response object in assignment
This represents the response from HMPPS Auth's '/api/authuser` endpoint,
which has a different structure to the usual User model we were
attempting to use before.

Naming these two interfaces was tricky, as they essentially refer to the
same thing, but I've opted to name them after the endpoints they're
returned by (User from '/api/user' and AuthUser from '/api/authuser').

What also made this tricky was that this the 'api/authuser' endpoints we
consume return an array of AuthUsers, rather than a single one. If this
were to return multiple users, we'd probably want to handle this better.