Mark referrals as assigned locally in mocks
I've spent hours trying to get this to work properly locally, but due to
the state transition, it's very tricky to mock the different stages of
assigning a referral and the different users being assigned.

I've decided to mark them as assigned so at least we can move onto the
next step of the user journey (the action plan).

Eventually, we'll be able to remove these mocks once we've properly
linked up the Interventions Service plumbing here, so I didn't think it
was worth spending any more time on.